:sparkles: feat: new transaction editing route

+++ server/src/api/v1/dto/summary_dto.go
@@ -1,7 +1,7 @@
 package dto
 
 type SummaryResponseDto struct {
-	Income  float64 `json:"income"`
-	Expense float64 `json:"expense"`
-	Balance float64 `json:"balance"`
+	Income  int64 `json:"income"`
+	Expense int64 `json:"expense"`
+	Balance int64 `json:"balance"`
 }
